+++ libai/data/__init__.py
@@ -14,9 +14,6 @@
 # limitations under the License.
 
 from .structures import DistTensorData, Instance
-from .bert_dataset import BertDataset
-from .gpt_dataset import GPT2Dataset
-from .t5_dataset import T5Dataset
 from .build import (
     build_image_train_loader,
     build_image_test_loader,
