+++ libai/data/samplers/__init__.py
@@ -13,9 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .cifar import CIFAR10Dataset, CIFAR100Dataset
-from .imagenet import ImageNetDataset
-from .mnist import MNISTDataset
-from .bert_dataset import BertDataset
-from .gpt_dataset import GPT2Dataset
-from .t5_dataset import T5Dataset
+from .samplers import CyclicSampler, SingleRoundSampler
